langchain_py/pyproject.toml
[project]
name = "langchain-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "jq>=1.11.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.0.0",
]
